finished JWT, AUTH, PRODUCTS AND USERS GUARDS

+++ src/products/products.service.ts
@@ -1,52 +1,66 @@
+import { UserRole } from './../users/user.role.enum';
+import { User } from './../users/users.model';
 import { CreateProductDto } from './dto/create-product.dto';
 import {Product} from './product.models'
-import {Injectable, NotFoundException} from '@nestjs/common'
+import {Injectable, NotFoundException, UnauthorizedException} from '@nestjs/common'
 import {InjectModel} from '@nestjs/mongoose';
 import {Model} from 'mongoose'
+import { GetUserId } from 'src/auth/get-user.decorator';
+import { GetUserRole } from 'src/auth/get-user-role.decrator';
 
 @Injectable()
 export class ProductsService{
     private products: Product[]=[];
     constructor(
-    @InjectModel('Product') private readonly productModel:Model<Product>
+    @InjectModel('Product') private readonly productModel:Model<Product>,
+    @InjectModel('User') private readonly userModel:Model<User>,
     ){}
 
-    async insertProduct(product:CreateProductDto):Promise<Product>{
-        const newProd=await this.productModel.create(product);
+    async insertProduct(product:CreateProductDto, @GetUserId() userId):Promise<Product>{
+        const newProd = await this.productModel.create(product);
+        console.log(userId);
+        newProd.creator = userId;
         await newProd.save();
-        return newProd.toObject();
+        return newProd;
     }
 
-    async getProducts():Promise<Product[]>{
-        return this.productModel.find().exec();
-       
+    async getProducts(@GetUserId() userId, @GetUserRole() role):Promise<Product[]>{
+        if (role==UserRole.USER) { return await this.productModel.find({creator:userId},{__v:0}).exec();}
+        else return this.productModel.find();
     }
 
-    async getSingleProduct(prodId:string):Promise<any>{
-        return this.findProduct(prodId);
-        
-    }
-
-    async updateProduct(prodId:string,changes:Partial<Product>):Promise<Product>{
-        const updatedProduct=await this.productModel.findOneAndUpdate({_id:prodId},changes);
-        updatedProduct.save();
-        return updatedProduct;
-    }
-
-    async deleteProduct(prodId:string):Promise<void>{
-        const result=await this.productModel.deleteOne({_id:prodId}).exec();
-        if(result.n===0){
-            throw new NotFoundException('No such a product here');
-        }
-    }
-    private async findProduct(id:string):Promise<Product>{
-        let product;
+    async getSingleProduct(prodId:string, @GetUserId() userId, @GetUserRole() role):Promise<any>{
         try{
-            product=await this.productModel.findById(id);
-        }
+            if (role === UserRole.USER) {
+                return await this.productModel.find({ _id: prodId, creator: userId });}
+            else {
+                return await this.productModel.findOne({ _id: prodId })}
+            }
         catch(error){
             throw new NotFoundException('No such product here');
         };
-        return product;
+    }
+
+    async updateProduct(prodId:string, changes:Product, @GetUserId() userId, @GetUserRole() role): Promise<Product>{
+        let updatedProduct;
+        if(role===UserRole.USER){
+            updatedProduct = await this.productModel.findOneAndUpdate({_id:prodId, creator: userId},changes);}
+        else{
+            updatedProduct = await this.productModel.findOneAndUpdate({_id:prodId}, changes);
+        }
+            updatedProduct.save();
+        return updatedProduct;
+    }
+
+    async deleteProduct(prodId:string, @GetUserRole() role):Promise<void>{
+        if(role===UserRole.ADMIN){
+            try{
+                await this.productModel.deleteOne({_id:prodId}).exec();
+            }
+            catch (error){
+                throw new NotFoundException('No such a product here');
+            }
+        }
+        else {throw new UnauthorizedException('Restricted rights')}
     }
 }
